Remove :skip_existence_check option
